pyproject.toml
[project]
name = "stock-analyzer-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "bs4>=0.0.2",
    "ipykernel>=7.1.0",
    "jupyter>=1.1.1",
    "langchain>=1.1.0",
    "langchain-openai>=1.1.0",
    "langgraph>=1.0.4",
    "langsmith>=0.4.53",
    "numpy>=2.3.5",
    "openai>=2.8.1",
    "pandas>=2.3.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "torch>=2.9.1",
    "transformers>=4.57.3",
    "yfinance>=0.2.66",
]
